+++ src/main/kotlin/dev/khulnasoft/intellij/Icons.kt
@@ -1,0 +1,9 @@
+package dev.khulnasoft.intellij
+
+import com.intellij.openapi.util.IconLoader
+
+class Icons {
+    companion object {
+        val Icon = IconLoader.getIcon("/icons/khulnasoft-no-padding.svg", this::class.java)
+    }
+}